Nested private SolutionCreator
